requirements.txt
Flask==2.3.3
groq==0.3.0
python-dotenv==1.0.0
pytest==7.4.0
